Move fetching logic to TemperatureEntry model
A lot of work here.  I set up reusable fetching methods to control the data coming from the API.  I also fixed the divs in the index to support 2 charts.
When entries are created from the API, the model will check if an entry matching that month, day and hour already exist.  If so, it will take the record high and low values from that entry, and delete the entry upon creation of the new entry.  This will alllow only a rolling year to exist, and won't break due to leap years not having Feb. 29th.

I'm still fetching the seed data manually, as it takes a very long time to test automating it, but I should be able to automate it by the time I deploy to Heroku.